cleaned up the module structure
